fix(gpu): Fix expand bench on multi-gpus
