+++ tfhe/src/integer/gpu/ciphertext/mod.rs
@@ -4,27 +4,15 @@ pub mod compressed_ciphertext_list;
 pub mod info;
 pub mod squashed_noise;
 
-use crate::core_crypto::gpu::lwe_bootstrap_key::{
-    prepare_cuda_ms_noise_reduction_key_ffi, CudaModulusSwitchNoiseReductionKey,
-};
 use crate::core_crypto::gpu::lwe_ciphertext_list::CudaLweCiphertextList;
 use crate::core_crypto::gpu::vec::CudaVec;
 use crate::core_crypto::gpu::CudaStreams;
-use crate::core_crypto::prelude::{
-    LweBskGroupingFactor, LweCiphertextList, LweCiphertextOwned, Numeric, UnsignedInteger,
-};
+use crate::core_crypto::prelude::{LweCiphertextList, LweCiphertextOwned};
 use crate::integer::gpu::ciphertext::info::{CudaBlockInfo, CudaRadixCiphertextInfo};
-use crate::integer::gpu::PBSType;
-use crate::integer::parameters::{
-    DecompositionBaseLog, DecompositionLevelCount, GlweDimension, LweDimension, PolynomialSize,
-};
+use crate::integer::parameters::LweDimension;
 use crate::integer::{IntegerCiphertext, RadixCiphertext, SignedRadixCiphertext};
-use crate::shortint::{CarryModulus, Ciphertext, EncryptionKeyChoice, MessageModulus};
+use crate::shortint::{Ciphertext, EncryptionKeyChoice};
 use crate::GpuIndex;
-use tfhe_cuda_backend::bindings::{
-    cleanup_expand_without_verification_64, cuda_expand_without_verification_64,
-    scratch_cuda_expand_without_verification_64,
-};
 
 pub trait CudaIntegerRadixCiphertext: Sized {
     const IS_SIGNED: bool;
@@ -526,101 +514,4 @@ impl From<EncryptionKeyChoice> for KsType {
             EncryptionKeyChoice::Small => Self::BigToSmall,
         }
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-/// # Safety
-///
-/// - `stream` __must__ be synchronized to guarantee computation has finished, and inputs must not
-///   be dropped until stream is synchronised
-///
-///
-/// In this method, the input `lwe_flattened_compact_array_in` represents a flattened compact list.
-/// Instead of receiving a `Vec<CompactCiphertextList>`, it takes a concatenation of all LWEs
-/// that were inside that vector of compact list. Handling the input this way removes the need
-/// to process multiple compact lists separately, simplifying GPU-based operations. The variable
-/// name `lwe_flattened_compact_array_in` makes this intent explicit.
-pub unsafe fn expand_async<T: UnsignedInteger, B: Numeric>(
-    streams: &CudaStreams,
-    lwe_array_out: &mut CudaLweCiphertextList<T>,
-    lwe_flattened_compact_array_in: &CudaVec<T>,
-    bootstrapping_key: &CudaVec<B>,
-    computing_ks_key: &CudaVec<T>,
-    casting_key: &CudaVec<T>,
-    message_modulus: MessageModulus,
-    carry_modulus: CarryModulus,
-    computing_glwe_dimension: GlweDimension,
-    computing_polynomial_size: PolynomialSize,
-    computing_lwe_dimension: LweDimension,
-    computing_ks_level: DecompositionLevelCount,
-    computing_ks_base_log: DecompositionBaseLog,
-    casting_input_lwe_dimension: LweDimension,
-    casting_output_lwe_dimension: LweDimension,
-    casting_ks_level: DecompositionLevelCount,
-    casting_ks_base_log: DecompositionBaseLog,
-    pbs_level: DecompositionLevelCount,
-    pbs_base_log: DecompositionBaseLog,
-    pbs_type: PBSType,
-    casting_key_type: KsType,
-    grouping_factor: LweBskGroupingFactor,
-    num_lwes_per_compact_list: &[u32],
-    is_boolean: &[bool],
-    noise_reduction_key: Option<&CudaModulusSwitchNoiseReductionKey>,
-) {
-    let ct_modulus = lwe_array_out.ciphertext_modulus().raw_modulus_float();
-    let mut mem_ptr: *mut i8 = std::ptr::null_mut();
-    let num_compact_lists = num_lwes_per_compact_list.len();
-
-    let ms_noise_reduction_key_ffi =
-        prepare_cuda_ms_noise_reduction_key_ffi(noise_reduction_key, ct_modulus);
-    let allocate_ms_noise_array = noise_reduction_key.is_some();
-
-    scratch_cuda_expand_without_verification_64(
-        streams.ptr.as_ptr(),
-        streams.gpu_indexes_ptr(),
-        streams.len() as u32,
-        std::ptr::addr_of_mut!(mem_ptr),
-        computing_glwe_dimension.0 as u32,
-        computing_polynomial_size.0 as u32,
-        computing_glwe_dimension
-            .to_equivalent_lwe_dimension(computing_polynomial_size)
-            .0 as u32,
-        computing_lwe_dimension.0 as u32,
-        computing_ks_level.0 as u32,
-        computing_ks_base_log.0 as u32,
-        casting_input_lwe_dimension.0 as u32,
-        casting_output_lwe_dimension.0 as u32,
-        casting_ks_level.0 as u32,
-        casting_ks_base_log.0 as u32,
-        pbs_level.0 as u32,
-        pbs_base_log.0 as u32,
-        grouping_factor.0 as u32,
-        num_lwes_per_compact_list.as_ptr(),
-        is_boolean.as_ptr(),
-        num_compact_lists as u32,
-        message_modulus.0 as u32,
-        carry_modulus.0 as u32,
-        pbs_type as u32,
-        casting_key_type as u32,
-        true,
-        allocate_ms_noise_array,
-    );
-    cuda_expand_without_verification_64(
-        streams.ptr.as_ptr(),
-        streams.gpu_indexes_ptr(),
-        streams.len() as u32,
-        lwe_array_out.0.d_vec.as_mut_c_ptr(0),
-        lwe_flattened_compact_array_in.as_c_ptr(0),
-        mem_ptr,
-        bootstrapping_key.ptr.as_ptr(),
-        computing_ks_key.ptr.as_ptr(),
-        casting_key.ptr.as_ptr(),
-        &raw const ms_noise_reduction_key_ffi,
-    );
-    cleanup_expand_without_verification_64(
-        streams.ptr.as_ptr(),
-        streams.gpu_indexes_ptr(),
-        streams.len() as u32,
-        std::ptr::addr_of_mut!(mem_ptr),
-    );
 }
